Clear gradient and hessian on shift
