+++ autode/opt/coordinates.py
@@ -51,11 +51,63 @@ class OptCoordinates(ValueArray, ABC):
         """Set the second derivatives of the energy"""
         self._h = value
 
-    @abstractmethod
     @property
+    @abstractmethod
     def type_str(self) -> str:
         """Type string of these coordinates. Should be compatible with to()"""
 
     @abstractmethod
     def to(self, *args, **kwargs):
         """Transformation between these coordinates and another type"""
+
+    def __setitem__(self, key, value):
+        """
+        Set an item or slice in these coordinates. Clears the current
+        gradient and Hessian as well as clearing setting the coordinates.
+        Does NOT check if the current value is close to the current, thus
+        the gradient and hessian shouldn't be cleared.
+        """
+
+        self.clear_gradient_and_hessian()
+        return super().__setitem__(key, value)
+
+    @abstractmethod
+    def _iadd(self, value: np.ndarray) -> 'OptCoordinates':
+        """Inplace addition of some coordinates"""
+
+    def __iadd__(self, other: np.ndarray):
+        """
+        Inplace addition of another set of coordinates. Clears the current
+        gradient vector and Hessian matrix.
+
+        Arguments:
+            other (np.ndarray): Array to add to the coordinates
+
+        Returns:
+            (autode.opt.coordinates.OptCoordinates): Shifted coordinates
+        """
+        self._iadd(other)
+        self.clear_gradient_and_hessian()
+        return self
+
+    def __isub__(self, other: np.ndarray):
+        """
+        Inplace subtraction of another set of coordinates. Clears the current
+        gradient vector and Hessian matrix.
+
+        Arguments:
+            other (np.ndarray): Array to subtract from the coordinates
+
+        Returns:
+            (autode.opt.coordinates.OptCoordinates): Shifted coordinates
+        """
+        return self.__iadd__(-other)
+
+    def clear_gradient_and_hessian(self) -> None:
+        """
+        Clear the gradient and Hessian for these coordinates. Called if the
+        coordinates have been perturbed, making these derivatives not
+        accurate any more for the new coordinates
+        """
+        self._g, self._h = None, None
+        return None
